factor out parseString to separate file XlsxString.h
