+++ src/XlsxCell.h
@@ -4,6 +4,7 @@
 #include <Rcpp.h>
 #include "rapidxml.h"
 #include "CellType.h"
+#include "XlsxString.h"
 
 // Key reference for understanding the structure of the XML is
 // ECMA-376 (http://www.ecma-international.org/publications/standards/Ecma-376.htm)
@@ -27,49 +28,6 @@ inline std::pair<int, int> parseRef(const char* ref) {
   }
 
   return std::make_pair(row - 1, col - 1); // zero indexed
-}
-
-// Parser for <si> and <is> inlineStr tags CT_Rst [p3893]
-// returns true if a string is found, false if missing.
-inline bool parseString(const rapidxml::xml_node<>* string, std::string *out) {
-  bool found = false;
-  out->clear();
-
-  const rapidxml::xml_node<>* t = string->first_node("t");
-  if (t != NULL) {
-    // According to the spec (CT_Rst, p3893) a single <t> element
-    // may coexist with zero or more <r> elements.
-    //
-    // However, software that read these files do not appear to exclusively
-    // follow this spec.
-    //
-    // MacOSX preview, considers only the <t> element if found and ignores any
-    // additional r elements.
-    //
-    //
-    // Excel 2010 appears to produce only elements with r elements or with a
-    // single t element and no mixtures. It will, however, consider an element
-    // with a single t element followed by one or more r elements as valid,
-    // concatenating the results. Any other combination of r and t elements
-    // is considered invalid.
-    //
-    // We read the <t> tag, if present, first, then concatenate any <r> tags.
-    // All Excel 2010 sheets will read correctly under this regime.
-    *out = std::string(t->value());
-    found = true;
-  }
-  // iterate over all r elements
-  for (const rapidxml::xml_node<>* r = string->first_node("r"); r != NULL;
-       r = r->next_sibling("r")) {
-    // a unique t element should be present (CT_RElt [p3893])
-    // but MacOSX preview just ignores chunks with no t element present
-    const rapidxml::xml_node<>* t = r->first_node("t");
-    if (t != NULL) {
-      *out += t->value();
-      found = true;
-    }
-  }
-  return found;
 }
 
 class XlsxCell {
