+++ src/XlsxString.h
@@ -1,0 +1,50 @@
+#ifndef READXL_XLSXSTRING_
+#define READXL_XLSXSTRING_
+
+#include <Rcpp.h>
+#include "rapidxml.h"
+
+// Parser for <si> and <is> inlineStr tags CT_Rst [p3893]
+// returns true if a string is found, false if missing.
+inline bool parseString(const rapidxml::xml_node<>* string, std::string *out) {
+  bool found = false;
+  out->clear();
+
+  const rapidxml::xml_node<>* t = string->first_node("t");
+  if (t != NULL) {
+    // According to the spec (CT_Rst, p3893) a single <t> element
+    // may coexist with zero or more <r> elements.
+    //
+    // However, software that read these files do not appear to exclusively
+    // follow this spec.
+    //
+    // MacOSX preview, considers only the <t> element if found and ignores any
+    // additional r elements.
+    //
+    //
+    // Excel 2010 appears to produce only elements with r elements or with a
+    // single t element and no mixtures. It will, however, consider an element
+    // with a single t element followed by one or more r elements as valid,
+    // concatenating the results. Any other combination of r and t elements
+    // is considered invalid.
+    //
+    // We read the <t> tag, if present, first, then concatenate any <r> tags.
+    // All Excel 2010 sheets will read correctly under this regime.
+    *out = std::string(t->value());
+    found = true;
+  }
+  // iterate over all r elements
+  for (const rapidxml::xml_node<>* r = string->first_node("r"); r != NULL;
+       r = r->next_sibling("r")) {
+    // a unique t element should be present (CT_RElt [p3893])
+    // but MacOSX preview just ignores chunks with no t element present
+    const rapidxml::xml_node<>* t = r->first_node("t");
+    if (t != NULL) {
+      *out += t->value();
+      found = true;
+    }
+  }
+  return found;
+}
+
+#endif
